Fixed circular imports in storage code
