+++ packages/server/src/admin/bot.ts
@@ -6,7 +6,7 @@ import { Readable } from 'stream';
 import { getConfig } from '../config/loader';
 import { getAuthenticatedContext } from '../context';
 import { Repository, getSystemRepo } from '../fhir/repo';
-import { getBinaryStorage } from '../fhir/storage';
+import { getBinaryStorage } from '../storage/loader';
 import { makeValidationMiddleware } from '../util/validator';
 
 export const createBotValidator = makeValidationMiddleware([
