+++ packages/server/src/cloud/aws/execute.test.ts
@@ -7,7 +7,7 @@ import express from 'express';
 import request from 'supertest';
 import { initApp, shutdownApp } from '../../app';
 import { getConfig, loadTestConfig } from '../../config/loader';
-import { getBinaryStorage } from '../../fhir/storage';
+import { getBinaryStorage } from '../../storage/loader';
 import { initTestAuth } from '../../test.setup';
 import { getLambdaFunctionName } from './execute';
