+++ packages/server/src/email/email.ts
@@ -5,7 +5,7 @@ import { sendEmailViaSes } from '../cloud/aws/email';
 import { getConfig } from '../config/loader';
 import { MedplumSmtpConfig } from '../config/types';
 import { Repository } from '../fhir/repo';
-import { getBinaryStorage } from '../fhir/storage';
+import { getBinaryStorage } from '../storage/loader';
 import { globalLogger } from '../logger';
 import { getFromAddress } from './utils';
