+++ packages/server/src/workers/index.test.ts
@@ -1,7 +1,7 @@
 import { closeWorkers, initWorkers } from '.';
 import { loadTestConfig } from '../config/loader';
 import { closeDatabase, initDatabase } from '../database';
-import { initBinaryStorage } from '../fhir/storage';
+import { initBinaryStorage } from '../storage/loader';
 import { loadStructureDefinitions } from '../fhir/structure';
 import { closeRedis, initRedis } from '../redis';
 import { seedDatabase } from '../seed';
